chore: move 'OngoingTransaction' to its own module.
